Keep namespace relative if base_prefix is empty string
Suppose that rather than give a different string for the base_prefix, a
user wishes to entirely remove the base prefix.  If the param
base_prefix is set to the empty string, then previously the topic names
would become absolute, i.e., would begin with "/".  This changeset
ensures that they remain relative in this case.